clean/backend/requirements.txt
fastapi
uvicorn
pinecone
ollama
python-dotenv